Network Commande Show

+++ resources/js/components/CommandeShow.vue
@@ -44,7 +44,8 @@ export default {
             sectionnable_type: false,
             list: false,
             label: '',
-            found: false
+            found: false,
+
         }
     },
     watch: {
@@ -225,8 +226,7 @@ export default {
                         product: this.selected_element,
                         type: 'App\\' + this.sectionnable_type
                     }
-                )
-                .then( response => {
+                ).then( response => {
                     // Grab la section
                     var section = this.commande.sections.find( (sect, section) => {
                         return sect.id ===  this.new_section
@@ -242,7 +242,7 @@ export default {
                         });
                         this.$forceUpdate()
 
-                        axios.get('https://azimuts.ga/article/api/changer-etat/' + this.selected_element.id + '/demandé').then(response => {
+                        axios.get('https://azimuts.ga/article/api/changer-etat/' + this.selected_element.id + '/wished').then(response => {
                             console.log(response.data);
 
                         }).catch(error => {
@@ -397,48 +397,66 @@ export default {
                 })
             }
         },
+        // Supprimer les produits d'une section
         deleteProductSection(section, article, type){
+            console.log('deleted')
+            this.$swal({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                if (result.isConfirmed){
+                    axios.get('/section-product/delete/' + article.id + '/' + section.id ).then(response => {
+                        console.log(response.data);
+                        if(response.data === 0){
+                            alert('Article Pas Supprimé. Veuillez Reesayé')
+                        } else {
+                            var section_trouvée = this.commande.sections.find(sect => {
+                                return sect.id === section.id
+                            })
+                            if(type === 'Article'){
+                                var article_trouvée = section_trouvée.articles.find( art => {
+                                    return art.id === article.id
+                                })
+                                var index = section_trouvée.articles.indexOf(article_trouvée)
+                                section_trouvée.articles.splice(index, 1)
+                                this.$forceUpdate()
 
-            axios.get('/section-product/delete/' + article.id + '/' + section.id ).then(response => {
-                console.log(response.data);
-                if(response.data === 0){
-                    alert('Article Pas Supprimé. Veuillez Reesayé')
-                } else {
-                    var section_trouvée = this.commande.sections.find(sect => {
-                        return sect.id === section.id
-                    })
-                    if(type === 'Article'){
-                        var article_trouvée = section_trouvée.articles.find( art => {
-                            return art.id === article.id
-                        })
-                        var index = section_trouvée.articles.indexOf(article_trouvée)
-                        section_trouvée.articles.splice(index, 1)
-                        this.$forceUpdate()
+                            } else if (type === 'Product'){
+                                var article_trouvée = section_trouvée.products.find( prod => {
+                                    return prod.id === article.id
+                                })
 
-                    } else if (type === 'Product'){
-                        var article_trouvée = section_trouvée.products.find( prod => {
-                            return prod.id === article.id
-                        })
-
-                        var index = section_trouvée.products.indexOf(article_trouvée)
-                        section_trouvée.products.splice(index, 1)
-                        this.$forceUpdate()
-                    }
+                                var index = section_trouvée.products.indexOf(article_trouvée)
+                                section_trouvée.products.splice(index, 1)
+                                this.$forceUpdate()
+                            }
 
 
-                    // alert('Article Suprrimé')
+                            // alert('Article Suprrimé')
+                        }
+                    }).catch(error => {
+                        console.log(error);
+                    });
                 }
-            }).catch(error => {
-                console.log(error);
-            });
+            })
+
         },
         saveQuantity(section, article){
-            console.log(article)
+            article.message = 'Sauvegarde en Cours...'
+            this.$forceUpdate()
             axios.put('/article-update',  {section : section, article: article}).then(response => {
                 console.log(response.data);
-
+                article.message = 'Sauvegarde Réussie.'
+                this.$forceUpdate()
             }).catch(error => {
                 console.log(error);
+                article.error = 'Sauvegarde Échouée. Veuillez vérifier votre connexion Internet'
+                this.$forceUpdate()
             });
         },
         openEditModal(section){
@@ -482,26 +500,43 @@ export default {
             });
         },
         removeProduct(section, produit, type){
-            axios.delete('/sectionnable/' + produit.id + '/' + section.id) .then(response => {
-                console.log(response.data)
-                if(type === 'Product'){
-                    var index = section.products.indexOf(produit)
-                    section.products.splice(index, 1)
-                } else {
-                    axios.get('https://azimuts.ga/article/api/changer-etat/' + produit.pivot.id + '/enregistré').then(response => {
-                        console.log(response.data);
-                        var index = section.articles.indexOf(produit)
-                        section.articles.splice(index, 1)
-                    }).catch(error => {
-                        console.log(error);
-                    });
+            this.$swal(
+                {
+                    title: 'Êtes-vous sûr de vouloir supprimer cette ressource?',
+                    text: "Cette action est irréversible!",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Oui, Supprimer!',
+                    cancelButtonText: 'Annuler'
+                }).then((result) => {
+                    if (result.value) {
+                        axios.delete('/sectionnable/' + produit.id + '/' + section.id) .then(response => {
+                            if(type === 'Product'){
+                                var index = section.products.indexOf(produit)
+                                section.products.splice(index, 1)
+                            } else {
+                                axios.get('https://azimuts.ga/article/api/changer-etat/' + produit.pivot.id + '/enregistré').then(response => {
+                                    console.log(response.data);
+                                    var index = section.articles.indexOf(produit)
+                                    section.articles.splice(index, 1)
+                                    this.$forceUpdate()
+                                }).catch(error => {
+                                    console.log(error);
+                                });
 
+                            }
+
+                            this.$forceUpdate()
+                            this.$swal('Produit Supprimé')
+                        }).catch(error => {
+                            console.log(error);
+                        });
+                    }
                 }
+            )
 
-                this.$forceUpdate()
-            }).catch(error => {
-                console.log(error);
-            });
         },
         majStock(){
             this.isLoading.majStock = true;
@@ -655,6 +690,12 @@ export default {
         }
         if(this.products_prop){
             this.products = this.products_prop
+            this.products.map( product => {
+                product.message = {
+                    text : '',
+                    color: ''
+                }
+            })
         }
 
         if(this.templates_prop){
@@ -692,6 +733,10 @@ export default {
                 })
             })
             this.articlesFetched.forEach( (article, index) => {
+                article.message = {
+                    text : '',
+                    color: ''
+                }
                 this.commande.sections.forEach( section => {
                     if( section.id === article.pivot.section_id ){
                         section.articles.push(article)
